Feature/exception 01 (#2)
* Criando estrutura básica

* Criando estrutura básica - application

* Criando JPA config

* Criando Exception

* Ajuste no delete

+++ service-schedule/src/main/java/com/serviceschedule/service/ServiceScheduleService.java
@@ -1,9 +1,14 @@
 package com.serviceschedule.service;
 
+
+import com.serviceschedule.exception.PrestadorDuplicadoException;
+import com.serviceschedule.exception.PrestadorNaoEncontradoException;
 import com.serviceschedule.model.ServiceScheduleModel;
 import com.serviceschedule.repository.ServiceScheduleRepository;
 import java.util.List;
 import java.util.Optional;
+import java.util.concurrent.ExecutionException;
+import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -27,10 +32,25 @@ public class ServiceScheduleService {
     }
 
     public ServiceScheduleModel savePrestador(ServiceScheduleModel serviceScheduleModel) {
+
+        String nomePrestador = serviceScheduleModel.getNome();
+
+        if (nomePrestadorJaCadastrado(nomePrestador)) {
+            throw new PrestadorDuplicadoException("Já existe um prestador com este nome cadastrado");
+        }
+
         return serviceScheduleRepository.save(serviceScheduleModel);
     }
 
-    public void deletePrestador(Long id) {
-        serviceScheduleRepository.deleteById(id);
+    public boolean nomePrestadorJaCadastrado(String nome) {
+        return serviceScheduleRepository.existsByNome(nome);
     }
-}
+
+    public void deletePrestador(Long id) {
+            if (!serviceScheduleRepository.existsById(id)) {
+                throw new PrestadorNaoEncontradoException("Prestador não encontrado com o ID: " + id);
+            }
+            serviceScheduleRepository.deleteById(id);
+        }
+      
+    }
